feat: create modal component and refactor settings to use it
