+++ src/components/Modal.css
@@ -1,0 +1,24 @@
+.settings {
+  position: absolute;
+  top: 0;
+  right: 0;
+}
+
+#portal-root .backdrop {
+  position: absolute;
+  background-color: var(--overlay-alpha);
+  z-index: 2;
+}
+
+#portal-root .modal-wrapper {
+  position: absolute;
+  padding: 1rem;
+  background-color: whitesmoke;
+  box-shadow: 0px 5px 15px 5px rgba(0, 0, 0, 0.08);
+  z-index: 5;
+  width: 20rem;
+}
+
+.backdrop .modal-wrapper {
+  pointer-events: none;
+}
